+++ planning/iris/BUILD.bazel
@@ -14,6 +14,7 @@ drake_cc_package_library(
     visibility = ["//visibility:public"],
     deps = [
         ":iris_from_clique_cover",
+        ":fast_iris"
     ],
 )
 
@@ -33,6 +34,24 @@ drake_cc_library(
         "//planning:visibility_graph",
     ],
     deps = [
+        "@common_robotics_utilities",
+    ],
+)
+
+drake_cc_library(
+    name = "fast_iris",
+    srcs = ["fast_iris.cc"],
+    hdrs = ["fast_iris.h"],
+    deps = [
+        "//planning:collision_checker",
+        "//common:parallelism",
+        "//geometry:meshcat",
+        "//geometry/optimization:convex_set",
+        "//solvers:choose_best_solver",
+        "//solvers:clarabel_solver",
+        "//solvers:mathematical_program",
+        "//solvers:mosek_solver",
+        "//solvers:solve",
         "@common_robotics_utilities",
     ],
 )
@@ -66,6 +85,29 @@ drake_cc_googletest(
         "//planning:scene_graph_collision_checker",
         "//planning/graph_algorithms:max_clique_solver_via_mip",
         "//systems/framework:diagram_builder",
+    ],
+)
+
+drake_cc_googletest(
+    name = "fast_iris_test",
+    timeout = "moderate",
+    data = ["//multibody/parsing:test_models"],
+    num_threads = 8,
+    # Most of these tests take an exceptionally long time under
+    # instrumentation, resulting in timeouts, and so are excluded.
+    tags = [
+        "no_asan",
+        "no_memcheck",
+    ],
+    deps = [
+        ":fast_iris",
+        "//planning:robot_diagram_builder",
+        "//planning:scene_graph_collision_checker",
+        "//common/test_utilities:expect_throws_message",
+        "//common/test_utilities:maybe_pause_for_user",
+        "//geometry:meshcat",
+        "//geometry/test_utilities:meshcat_environment",
+        "//multibody/inverse_kinematics",
     ],
 )
