+++ planning/iris/fast_iris.cc
@@ -1,4 +1,4 @@
-#include "drake/planning/fast_iris.h"
+#include "drake/planning/iris/fast_iris.h"
 
 #include <algorithm>
 #include <iostream>
